Apply old Pin Descriptor doc fix from emc2.git
The change in this commit came from the emc2.git repo.

The 3x20 has so many IO pins that the section of the hostmot2 register
file devoted to Pin Descriptors filled up, and there was no longer space
for the "end of list" sentinel.  This commit documents this new behavior.

Also, removes a control-character topy.