build: make a symlink to settings.sh do the trick
in ise 12 on ubuntu, sourcing settings.sh breaks evince(!) so doing it
in bashrc is a poor idea.  remembering to do it before each build is a
pain.  so, offer a third method: a symlink settings.sh -> ..../settings.sh

when this symlink is present, every command run by build.py first sources
the file and then executes the command.