turn on register ordering
The syntax of the -r option to the map command changed between 10.1
and 13.3.  The -r option in 10.1 is a flag (accepts no option-argument),
while the -r option in 13.3 requires an argument.

In 10.1, no -r (the default) means "do register ordering" and with -r
it means "don't do register ordering".  In 13.3, no -r (the default)
means "do register ordering with 4 (or sometimes 8) registers per slice"
(whatever that means).  In any case, no -r on both means "do register
ordering".

Before this commit, the build system calls map with the 10.1-compatible
"-r" command-line syntax to disable register ordering, but this syntax
doesn't work with 13.3 (the 13.3 syntax to do the same thing is "-r off").

Peter Wallace says:
> I would be tempted to just leave it off altogether for 13.3. I think
> its was just needed for one or two 5I20/4I65 configs to fit. The -r
> disables register ordering which wastes a tiny bit of space (say .1%).
>
> But really any config thats that close to the edge should probably
> be dropped/trimmed...