deb: use dh for binary-indep and clean targets
"dh $@" in the recipe for a debian/rules target runs the appropriate
sequence of debhelper commands for that target.

This does some things we don't need, but it also doesn't skip any of the
things we do need, that the hand-written recipe missed, like dh_fixperms
(fixed by hand in a previous commit).

The only trickiness here is overriding dh_installdocs to include the
README and the regmap files.

"dh $@" does run dh_testroot when needed, so i got rid of the manual
check-root target.